requirements.txt
marshmallow==3.5.1
pynsq==0.8.3  